refactor(cli): extract shared operator readiness reporting
The decap, forward, pipeline and route operator CLIs each carried a nearly
identical copy of the readiness table rendering and summary logic. Move it
into a shared module in the CLI core library, leaving each operator with
only its own command definitions and RPC wiring.

No behavior change: the rendered output, exit code semantics and the JSON
payload are all unchanged.
