+++ cli/core/src/lib.rs
@@ -9,6 +9,7 @@ pub mod errors;
 pub mod humanfmt;
 pub mod logging;
 pub mod output;
+pub mod readiness;
 
 /// Initialise the logger and output backend from a `Format` choice.
 ///
