+++ operators/pipeline/cli/Cargo.toml
@@ -14,10 +14,8 @@ clap = { version = "4.5", features = ["derive"] }
 clap_complete = { version = "4.5", features = ["unstable-dynamic"] }
 prost = "0.13"
 log = "0.4"
-colored = "3"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-tabled = { version = "0.18", features = ["ansi"] }
 tokio = { version = "1", features = ["rt", "net", "time", "macros", "sync"] }
 tonic = { version = "0.13", features = ["gzip"] }
